+++ ddf_utils/chef/ingredient.py
@@ -3,6 +3,9 @@
 """main ingredient class"""
 
 import pandas as pd
+import numpy as np
+from ..str import format_float_digits
+from .helpers import read_opt
 import os
 import logging
 
@@ -55,11 +58,98 @@ class BaseIngredient(object):
         self.data = None
         return
 
+    def serve(self, outpath, **options):
+        """save the ingledient to disk.
+
+        Parameters
+        ----------
+        outpath : str
+            the path to save the ingredient
+
+        Other Parameters
+        ----------------
+        digits : int
+            how many digits to keep at most.
+
+        """
+        data = self.copy_data()
+        t = self.dtype
+        assert isinstance(data, dict)
+        for k, df in data.items():
+            # change boolean into string
+            for i, v in df.dtypes.iteritems():
+                if v == 'bool':
+                    df[i] = df[i].map(lambda x: str(x).upper())
+            if t == 'datapoints':
+                by = self.key_to_list()
+                path = os.path.join(outpath, 'ddf--{}--{}--by--{}.csv'.format(t, k, '--'.join(by)))
+            elif t == 'concepts':
+                path = os.path.join(outpath, 'ddf--{}.csv'.format(t))
+            elif t == 'entities':
+                domain = self.key[0]
+                if k == domain:
+                    path = os.path.join(outpath, 'ddf--{}--{}.csv'.format(t, k))
+                else:
+                    path = os.path.join(outpath, 'ddf--{}--{}--{}.csv'.format(t, domain, k))
+            else:
+                raise ValueError('Not a correct collection: ' + t)
+            # formatting numbers for datapoints
+            if t == 'datapoints':
+                digits = read_opt(options, 'digits', default=5)
+                df = df.set_index(by)
+                if not np.issubdtype(df[k].dtype, np.number):
+                    try:
+                        df[k] = df[k].astype(float)
+                        df[k] = df[k].map(lambda x: format_float_digits(x, digits))
+                    except ValueError:
+                        logging.warning("data not numeric: " + k)
+                else:
+                    df[k] = df[k].map(lambda x: format_float_digits(x, digits))
+                df[[k]].to_csv(path, encoding='utf8')
+            else:
+                df.to_csv(path, index=False, encoding='utf8')
+
 
 class Ingredient(BaseIngredient):
     """
     ingredient class: represents an ingredient object in recipe file.
     see the implement of from_dict() method for how the object is constructed.
+
+    Here is an example ingredient object in recipe:
+
+    .. code-block:: yaml
+
+        id: example-ingredient
+        dataset: ddf--example--dataset
+        key: "geo,time"  # key columns of ingredient
+        value:  # only include concepts listed here
+          - concept_1
+          - concept_2
+        filter:  # select rows by column values
+          geo:  # only keep datapoint where `geo` is in [swe, usa, chn]
+            - swe
+            - usa
+            - chn
+
+    Attributes
+    ----------
+    ingred_id : `str`
+        The id string
+    ddf_id : `str`
+        the underlaying dataset id
+    key : `str`
+        the key columns of the ingredient
+    value : `list`
+        concept filter applied to the dataset. if `value` == "*", all concept of
+        the dataset will be in the ingredient
+    row_filter : `dict`
+        row filter applied to the dataset
+
+    Methods
+    -------
+    get_data()
+        read in and return the ingredient data
+
     """
     def __init__(self, ingred_id,
                  ddf_id=None, key=None, values=None, row_filter=None, data=None):
@@ -71,6 +161,16 @@ class Ingredient(BaseIngredient):
 
     @classmethod
     def from_dict(cls, data):
+        """create an instance by a dictionary
+
+        The dictionary should provide following keys:
+
+        - id
+        - dataset
+        - key
+        - value
+        - filter (optional)
+        """
         ingred_id = data['id']
         ddf_id = data['dataset']
         key = data['key']
@@ -84,6 +184,7 @@ class Ingredient(BaseIngredient):
 
     @property
     def ddf(self):
+        """The DDF reader object"""
         if self._ddf:
             return self._ddf
         else:
@@ -94,6 +195,7 @@ class Ingredient(BaseIngredient):
 
     @property
     def ddf_path(self):
+        """The path to the dataset"""
         return self.ddf.dataset_path
 
     def __repr__(self):
@@ -148,31 +250,34 @@ class Ingredient(BaseIngredient):
         else:
             return {'concepts': self.ddf.get_concepts()[self.values]}
 
-    def filter_row(self, df):
-        """return the rows selected by self.row_filter."""
-        # TODO: improve filtering function
-        # 1. know more about the row_filter syntax
-        # 2. The query() Method is Experimental
-        if self.row_filter:
-            query_str = "and".join(["{} in {}".format(k, v) for k, v in self.row_filter.items()])
-            # print(query_str)
-            df = df.query(query_str)
-
-        return df
 
     def get_data(self, copy=False, key_as_index=False):
+        """read in and return the ingredient data
+        """
         funcs = {
             'datapoints': self._get_data_datapoint,
             'entities': self._get_data_entities,
             'concepts': self._get_data_concepts
         }
+
+        def filter_row(df):
+            """return the rows selected by self.row_filter."""
+            # TODO: improve filtering function
+            # 1. know more about the row_filter syntax
+            # 2. The query() Method is Experimental
+            if self.row_filter:
+                query_str = "and".join(["{} in {}".format(k, v) for k, v in self.row_filter.items()])
+                # print(query_str)
+                df = df.query(query_str)
+            return df
+
         if self.data is None:
             data = funcs[self.dtype](copy)
             for k, v in data.items():
                 if self.row_filter:
                     index_cols = data[k].index.names
                     # data[k] = self.filter_row(data[k].reset_index()).set_index(index_cols)
-                    data[k] = self.filter_row(data[k].reset_index())
+                    data[k] = filter_row(data[k].reset_index())
                 else:
                     data[k] = data[k].reset_index()
             self.data = data